# ProyectoFinal/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Projecte Final ",
        "description": "REST API. Projecte Final. DAW Client i servidor Joan R, Dimitry C, Javier P.",
        "contact": {
            "name": "Joan Ramis.",
            "email": "joanramis1@paucasesnovescifp.cat"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://www.rampacom.com/ProyectoFinal/public/"
        },
        {
            "url": "http://127.0.0.1/projecte/TreballFinal/ProyectoFinal/public/"
        }
    ],
    "paths": {
...
        "/api/municipio/{id}": {
            "get": {
                "tags": [
                    "Municipios"
                ],
                "summary": "Mostrar un Municipio por ID",
                "description": "Descripcion de un Idioma.",
                "operationId": "6f10de88309d4abcbf330e1a491dcee5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del Municipio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "id": {
                                "summary": "Introduce el numero de ID del Municipio",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informacion del Municipio.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Hay un error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Municipio no encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/municipio/borra/{id}": {
            "delete": {
                "tags": [
                    "Municipios"
                ],
                "summary": "Borra un Municipio",
                "description": "Borra un Municipio. Solo por Administradores",
                "operationId": "a02d2c6e68eae5746763403bc1f53d3b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Municipio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "ID no encotrada"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/municipio/crea": {
            "post": {
                "tags": [
                    "Municipios"
                ],
                "summary": "Crea un Municipio",
                "description": "Crea un Municipio. Solo por Administradores.",
                "operationId": "0eec2fdc898a9600c60248e3d3f01cbc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Esto es un nuevo Municipio"
                                    },
                                    "isla": {
                                        "type": "enum",
                                        "format": "enum",
                                        "example": "Esto es la isla del Municipio"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atributo obligatorio requerido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/municipio/modifica/{id}": {
            "put": {
                "tags": [
                    "Municipios"
                ],
                "summary": "Modifica un Municipio",
                "description": "Modifica un Municipio. Solo por Administradores.",
                "operationId": "a0ecf94f99f964b5b0841cf5096b1603",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id Municipio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Esto es un nuevo Municipio"
                                    },
                                    "islas": {
                                        "type": "enum",
                                        "format": "enum",
                                        "example": "Esto es la isla del Municipio"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": "error"
                                        },
                                        "data": {
                                            "type": "string",
                                            "example": "Atributo obligatorio requerido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}